refactor: interface adjustments
