--- src/electron/data/protocols/I18n/index.ts
@@ -1,10 +1,0 @@
-export * from './I18nLoad';
-export * from './I18nLoadAdditional';
-export * from './I18nGetLanguages';
-export * from './I18nTranslate';
-export * from './I18nCurrentLanguage';
-export * from './I18nChangeLanguage';
-export * from './I18nInitialize';
-export * from './I18nOnLoaded';
-export * from './I18nOnLanguageChanged';
-export * from './I18nGetResource';
