+++ src/electron/data/protocols/Resource/Resource.ts
@@ -1,0 +1,14 @@
+import { ResourceGetDatabasePath } from './ResourceGetDatabasePath';
+import { ResourceGetIcon } from './ResourceGetIcon';
+import { ResourceGetLanguages } from './ResourceGetLanguages';
+import { ResourceGetMigrationPath } from './ResourceGetMigrationsPath';
+import { ResourceGetSelectedLanguages } from './ResourceGetSelectedLanguage';
+import { ResourceWriteLanguageFile } from './ResourceWriteLanguageFile';
+
+export default interface Resource
+  extends ResourceGetDatabasePath,
+    ResourceGetIcon,
+    ResourceGetLanguages,
+    ResourceGetMigrationPath,
+    ResourceGetSelectedLanguages,
+    ResourceWriteLanguageFile {}
