--- src/electron/data/protocols/Resource/index.ts
@@ -1,6 +1,0 @@
-export * from './ResourceGetSelectedLanguage';
-export * from './ResourceGetLanguages';
-export * from './ResourceGetIcon';
-export * from './ResourceWriteLanguageFile';
-export * from './ResourceGetDatabasePath';
-export * from './ResourceGetMigrationsPath';
