+++ src/electron/data/protocols/index.ts
@@ -1,5 +1,3 @@
-export * from './I18n';
 export * from './Menu';
-export * from './Resource';
 export * from './db';
 export * from './listen';
